refactor(ui): extract AddEndpointModal and UploadStrip to stay under 1000-line file-size gate
